meta-sp7/conf : amd flash layout rofs and rwfs update
Increased the rofs partition size to 77.5MB and reduced
the rwfs partition size to 32MB. Without this change, BMC
is rebooting three times on Morocco and Congo. With this
change that number changed to one. And at times on Congo
board there is no reboot.

Following is the new BMC SPI Flash Layout:
     - uboot     0x00      (2MB)
     - uboot-env 0x200000  (128KB)
     - kernel    0x220000  (16MB)
     - rofs      0x1220000 (77.875MB)
     - rwfs      0x6000000 (32MB)

Tested: verfied on Mrocco and Congo boards. Following tests were
        performed.
        - BMC booted successfully after flashing with RPI/dpcmd
	- did multiple power cycles after the update and BMC
	  booted successfully in all the cycles
	- updated the BMC firmware successfully from the webui

Signed-off-by: Mahesh Kurapati <mahesh.kurapati@amd.com>
Change-Id: I5f4199ac4bb4bfd5b473f6c17c23f9298e59370a