meta-amd:common:dbus Multi-host interface
This commit introduces a Multi-host interface that enables setting
and retrieval of configuration parameters needed to enable multi-host
infrastructure.

Note: This change is kept in the meta-common layer to avoid
overhead in patch management and to ensure broader applicability.
Future references should consider the benefits of maintaining such
changes in common layer to facilitate easier maintenance and updates.

Signed-off-by: aasboddu <Aashray.Boddu@amd.com>
Change-Id: I1424448a2e609c89664677059ce594f6d0df1c0d